Add explanation for `smc_checksum`
tbh, I don't understand why this is needed at all.  The documentation says "drag it to wherever you want (e.g. the Application Folder). ", yet it seems that the checksum requires it be at a certain path.
When I try compiling and running this on my own I get a different hash and the program won't run.  I either need to change `smc_checksum` each time I move the program or just remove 249:252.  After getting bored of changing `smc_checksum` I did the latter and the program works flawlessly now.

Am I correct,  or am I misunderstanding the role of `smc_checksum`?

Thanks for the fork!  Was disappointed the upstream wasn't working, and was relieved to find your fork.

Would it be okay to just go ahead and remove 249:252?